pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "polygon-helper"
version = "0.1.0"
description = "Tiny helper functions for Polygon.io pipelines"
readme = "README.md"
requires-python = ">=3.9"
authors = [{name = "Matte762"}]
dependencies = ["requests>=2.32.0", "pandas>=2.0.0", "python-dateutil>=2.8.2", "tenacity>=8.2.3"]

[project.urls]
Homepage = "https://github.com/your-username/polygon-helper"

[tool.setuptools.packages.find]
where = ["src"]
